# chore: Add benchmarks that use dedicated optimization path by System.Linq